Desplegando Mejoras Alpha Test v8 manejo del cambio

+++ acq/views.py
@@ -1,11 +1,9 @@
 from django.shortcuts import render
 from django.http import HttpResponse, HttpResponseRedirect
-import socket
 import json
-
 from django.http import JsonResponse
 from .models import Tag, Tk, PatioTanque,Tct
-from datetime import timedelta, datetime
+#from datetime import timedelta, datetime
 from django.template.response import TemplateResponse
 
 
@@ -31,16 +29,13 @@ from django.shortcuts import render,  get_object_or_404
 
 import csv
 from django.core.validators import DecimalValidator, ValidationError
-from django.core.exceptions import ValidationError, ValidationError
+from django.core.exceptions import ValidationError
 import pandas as pd
 from .validaciones import validar_parametro_tct as valida
-from django.utils import timezone
-
-
 
 def actualizar(request):
 
-   with open ('/home/morenomx/solucionesweb/sacvc/datos.json', encoding='utf-8') as data_file: #abre un archivo json
+   with open ('/home/morenomx/solucionesweb/sacvc/datos.json', encoding='utf-8') as data_file: # OJO MEJORAR
       dataf = json.loads(data_file.read())
       data_file.close()
 
@@ -195,19 +190,18 @@ def integridad_TCT(request, pk):
 
   DataFrame=pd.read_csv(file, delimiter='\t', ) #abre el csv tc y lo pasa a un dataframe
 
-  json_temp = {}
+
   json_temp = []
 
 
-  register_range=[]
+  #register_range=[]
 
   for i in range(0, len(DataFrame)):
 
-    register_range.append(i)
+    #register_range.append(i)
     nivel_format=format(DataFrame.iloc[i]['nivel']).replace(',','.')
     volumen_format=format(DataFrame.iloc[i]['volumen']).replace(',','.')
-    print(nivel_format)
-    print(volumen_format)
+
 
     nivel=valida(float(nivel_format),nivel_minimo,nivel_maximo)
     volumen=valida(float(volumen_format),volumen_minimo,volumen_maximo)
@@ -222,18 +216,20 @@ def integridad_TCT(request, pk):
 
 def guardar_TCT_BD(request, pk):
 
-
   try:
     #obj1 = Tct.objects.get(pk=pk)
     obj_tk =  Tk.objects.get(pk=pk)
-    Tct.objects.all().delete()
 
-    file=obj_tk.tct_archivo.path
-    DataFrame=pd.read_csv(file, delimiter='\t', ) #abre el csv tc y lo pasa a un dataframe
-    #nivel=DataFrame.iloc[3]['nivel']
-
-    for i in range(0, len(DataFrame)):
-     Tct.objects.create(id=None,  Lt0=DataFrame.iloc[i]['nivel'], Tov0=DataFrame.iloc[i]['volumen'], id_tk=obj_tk)
+    if obj_tk.tctvalido:
+        Tct.objects.all().delete()
+        file=obj_tk.tct_archivo.path
+        DataFrame=pd.read_csv(file, delimiter='\t', ) #abre el csv tc y lo pasa a un dataframe
+        for i in range(0, len(DataFrame)):
+            nivel_format=format(DataFrame.iloc[i]['nivel']).replace(',','.')
+            volumen_format=format(DataFrame.iloc[i]['volumen']).replace(',','.')
+            nivel=float(nivel_format)
+            volumen=float(volumen_format)
+            Tct.objects.create(id=None,  Lt0=nivel, Tov0=volumen, id_tk=obj_tk)
